fix(stats): blog injector + REST endpoint share public/exclude rules
PR #281's GET /web/stats/panels/:id/data added the metadata.public gate
+ server-side exclude_columns strip. But the OTHER path that exposes
panel data to non-admin consumers — `injectStatsPanelData`, called when
the storefront fetches a blog post — had neither guard. So a blog
editor (or compromised one) could embed `<statsPanel panelId="…">` in
TipTap content, the backend would resolve ANY referenced panel, and
the data shipped to the storefront uncensored.

Fix
- New `modules/stats/public-utils.ts` with `isPanelPublic(panel)` and
  `stripExcludedColumns(display, data)` so both consumers (blog inject,
  REST endpoint) call the SAME helpers — no rule drift.
- `injectStatsPanelData` now skips resolution for non-public panels
  and emits `{ error: 'panel_not_public', data: null }` so the
  storefront renders an empty state instead of leaking.
- Same helper applies `exclude_columns` strip so joined entities (e.g.
  the `inventory_item` join on a raw_materials panel) never reach the
  storefront if the editor opted them out.
- Public REST endpoint refactored to call the same helpers — no
  behavior change, just less duplication.

How a panel becomes embeddable in a blog:
  PATCH /admin/stats/panels/<id>
  { "metadata": { "public": true } }

Existing TipTap StatsPanelExtension on jyt-web continues to work
unchanged — it already reads from pre-injected node.attrs and renders;
the difference is the injector now sets `data: null` for non-public
panels so the renderer shows its empty/error state.
